fix confd wait issue
when trying to boot up the registry, it would skip setting etcd
values if anything was set. Because we made some changes to the
cookbook to set some variables to configure the storage backend,
this broke the container. Changing it to set the secretKey if
not present by default fixes the original intention of this
logic.