Let berkshelf resolve dependencies itself
Currently, the depenencies for the deis cookbook are listed
both in the Berksfile in the deis project and the metadata for
the deis cookbook. They should only live in the deis cookbook,
and dependencies will be automatically resolved when Berkshelf
pulls that cookbook.

This can cause issues, as the deis cookbook locks its dependencies
to known versions and the deis project Berksfile just pulls the
latest versions.

We don't need a Berksfile.lock anymore either, as the cookbook
now pins its depencies properly.