docs(standards): add commit style guide
this introduces a new commit standard for Deis developers. The
goal of this commit is to allow developers to search through
the history better, provide us better information through
commit messages, and eventually allowing us to script the
CHANGELOG.md so that we do not have to rely on a third-party
tool to do the parsing for us.