fix(controller): use postgresql for tests
sqlite3 does not support database transactions, which Deis uses
heavily in both the scheduler as well as in other components.
Regardless, our dev and test environments should be running the
tests on the same environment, so that we can ensure that the tests
should run the same in both environments.