refactor(discovery): remove discovery component
This component was used to provide etcd onto a host. Now that we are
using CoreOS as our base distribution which has etcd built in, we no
longer require this component. Going forward, if we decide to run
Deis on a different distribution, we will install etcd on the host.