refactor(contrib): update Deis deployment on AWS
The old contrib scripts used to assist with distributing an AMI with
Deis preconfigured for both the controller and for the nodes across
Amazon EC2 regions. It also helped with spawning a node on EC2
using the controller AMI, then bootstrapped the node so that the
node would be added to Chef. The AMI was meant to help speed up
the bootstrap process, as it pulled the latest Deis components
onto the AMI as well as the Docker daemon along with a couple
other packages.

With the shift to the scheduler, we no longer manage the difference
between controller, proxy and runtime nodes. We let
https://github.com/coreos/fleet do the heavy lifting for us. Instead,
we communicate with a CoreOS cluster and upload a couple of systemd
units onto the cluster. With that shift in thought, we can leverage
Amazon CloudFormation to bootstrap a CoreOS cluster and then do the
rest remotely with fleetctl.