fix(logger): fix publish loop
before, the logger's publish loop would be searching for a TCP
connection. Since the logger runs on UDP, the publish loop would
not find it running and would run forever. This changes netstat
to search over UDP for the logger instead, allowing it to publish
its values to etcd.

fixes #738