fix(controller): fix registry API calls
After upgrading the private registry to 0.6.8, there were some
changes to the API that were missed. For example, the docker version
is now mandatory[1].

If you have docker v0.10.0 or higher, there is also a checksum computation
check done on each push, so I hardcoded the user-agent to v0.9.0. This is
a stopgap until

a) we update the API calls to v0.10.0
b) we replace this component with a docker-registry python package[2]

[1]: https://github.com/dotcloud/docker-registry/commit/64f4198c9041c25ac416b8ebba5c57304315156c
[2]: https://github.com/dotcloud/docker-registry/pull/297

fixes #779