refactor(seed-deis-registry): remove seed-deis-registry dependency
Currently, startup takes forever because registry blocks until the
slugrunner container has been pulled. Users think things have stalled
and/or are broken - this is a bad user experience.

The ExecStartPost of the registry service already pulls the slugrunner
image, so we don't need to block until this job occurs beforehand.
Additionally, it's friendlier if the controller blocks until registry
and logger are ready - if a user attempts to make an account, all of
Deis should be up.