refactor(Makefile): better multi-node support
Our READMEs and Makefile really fall short for multi-node setups.
Specifically, we allude to the fact that DNS records will need to be
setup, but we don't provide additional info.

This commit updates the main README, contrib/ec2, and contrib/rackspace
with specific instructions for accessing the controller and router
for multi-node setups.

Additionally, the Makefile has been refactored so that there aren't
separate paths for 1-node and multi-node setups. Now we always start
the router, as we have provided instructions regarding DNS in the main README.