fix(test): specify machine IPs
The Rakefile breaks for cloud providers, as we used the HOST
for both the cluster domain and a list of hosts. Since etcd/fleet
now listen exclusively on the `private_ipv4` interface, connecting
on the public domain doesn't work.

This commit adds a new environment variable, `DEIS_TEST_HOSTS`,
to pass in the hosts to the creation of the cluster.