fix(registry): store apps as appname
Storing the image in the registry as user/appname introduces a couple
of confusing ideas:

- images are implied to be unique across accounts
- images are owned by one user and one only

Point 1 is incorrect. App names are completely unique across the
entire cluster[1]. Point 2 is incorrect as well because apps can
be shared with the `deis perms` command. Changing the name of the
image in the registry from `user/appname` to `appname` makes the
image names unique across the entire cluster as well as making the
images more easily sharable between users.

fixes #729

[1]: https://github.com/deis/deis/blob/d1e14a6f76789a86f1959ba7bf6057b3abe3b960/controller/api/models.py#L109

BREAKING CHANGES: because this changes the way that applications
are stored in the registry, this is a backwards-incompatible change
with applications prior to this commit. To migrate over to this change,
you'll have to destroy and re-create your apps with the same name
as it was before. The docker layers from previous builds will still
exist in the registry, but the repository names will be changed
to reflect this change.