fix(controller): remove username length validation
This fix that was implemented in #631 was a bit hackish, and was
only used because we stored the images as username/appname in the
registry. The registry has a restriction where the username has
to be greater than or equal to 4 in length. Now that we are not
storing the images with the associated owner's username, we no
longer need this restriction.

reverts #631