feat(contrib/coreos/user-data): create Deis motd
With the upgrade to CoreOS 310.0.0 in #911, the CoreOS motd
lists failed units, 3 of which we expect to be failed (since
we are stopping the reboot manager service). It probably
makes sense for us to manage our own MOTD anyway, with Deis
support information.

This commit works for both CoreOS 298.0.0 and CoreOS 310.0.0.