feat(Vagrantfile): add fallback to rsync
In some cases (namely when using an encrypted filesystem),
a `vagrant up` will fail due to an NFS error. In these cases,
it is necessary to fall back to rsync to share code between the
local filesystem and the VM.

This commit adds a comment to the Vagrantfile with these instructions
and the necessary config directive.

related to #942