fix(Makefile): use fleetctl to filter actual deis-* units
Call `fleetctl` to see which deis-* units are actually loaded, and then
reference only those in the `make stop` and `make uninstall` commands.
This avoids requiring special logic for router(s) except on start.
Requires `awk`, `sed`, and `tr` utilities, which should be ubiquitous.

This also is future-proofing for `fleetctl` 3.x, which returns errors in
many cases where version 2.0 allowed us to be sloppy. Thus the `deis_units`
def is written to accept filters on columns 2 (LOAD) and 4 (SUB) of the
fleetctl output. We will make more use of this when moving to CoreOS 317.0.0.

Closes #991.