feat(builder): build apps from more than master
Previously, the builder would completely disregard any pushes from
a branch other than master. This moves the builder command into
a post-receive hook, such that we can reference the branch that
is being pushed to the repository and send that to the builder.

For some reason, this PR also gives us Heroku compatibility on
issue #736, so this PR kills two birds with one stone.

TESTING: To test, we need to successfully push an app from a branch
other than master. To do that, run the following:

    $ make -C builder build restart
    $ cd /tmp
    $ git clone https://github.com/opdemand/example-perl && cd example-perl
    $ git checkout -b test-branch
    $ deis create
    $ git push deis test-branch

To test for the #736 ninja fix, try pushing again. You should see a
"Everything up-to-date" message displayed on your screen, and no
push should occur.

fixes #975, #736