fix(Makefile): Information about missing FLEETCTL_TUNNEL
For the following commands, it blows up in a non-informational way.

```
09:24 deis‹master› » git rev-parse HEAD
33785320c245dd6ba9285eca0e5a85d122318dc0
```

```
09:26 deis‹master› » make --version
GNU Make 3.81
Copyright (C) 2006  Free Software Foundation, Inc.
This is free software; see the source for copying conditions.
There is NO warranty; not even for MERCHANTABILITY or FITNESS FOR A
PARTICULAR PURPOSE.

This program built for i386-apple-darwin11.3.0
```

```
09:24 deis‹master› » make pull
Makefile:10: *** commands commence before first target.  Stop.
```

After this fix:

```
09:37 deis‹master› » make pull
Makefile:10: *** You need to set FLEETCTL_TUNNEL to the IP address of a
server in the cluster..  Stop.
```