fix(Makefile): enable targets to work for non-Vagrant
Adds DEIS_HOSTS, which is set automatically or can be overridden
for non-Vagrant platforms. This is used by ssh_all to determine
which hosts to SSH into - previously, we used `vagrant ssh`,
which obviously only works on Vagrant.

This also updates all the Makefiles to use fleet as appropriate,
and to pull / build / etc. on all DEIS_HOSTS.

TESTING: try the various make commands on Vagrant and Rackspace/EC2

replaces #998
closes #811