feat(client): add config:pull command
This command pulls in existing config variables set with deis
config:set and writes them out to a local .env file for
development purposes.

If a .env file exists and a environment variable in that file is
present in the application's config, the file's envvar will be
chosen unless the --overwrite option is selected. If
--interactive is present, a prompt to write the environment
variable to the file will be shown.

fixes #560