feat(router): proxy builder through router
The router component is to be the point of entry for the entire Deis cluster.
This commit makes the router forward connections on port 2222 to the builder.

This requires manually compiling nginx from source with the tcp_proxy module
and adding an appropriate upstream for builder.

TESTING: rebuild the router and builder (both need to be resubmitted with fleet)
```console
$ make -C router uninstall build run
$ make -C builder uninstall build run
```

And focus testing in two areas:
* That nothing was broken with the change to custom-built nginx (use Deis)
* That the router is properly proxying git/SSH (do lots of `git push`es)

closes #535