fix(router): increase timeouts and survive controller/builder failures
Bumps the timeouts to 20 minutes and wraps the builder and controller
proxies in `if` clauses to prevent the router from crashing when
either controller or builder are restarted.

Also adds a generic health-check URL to be used for load balancers,
and an X-Deis-Upstream header.

TESTIING: rebuild and restart the router:

```console
	$ make -C router/ build restart
```

closes #1096, #1098, #1078, #1081
replaces #1080, #1057