fix(registry): use confd templated config
A regression from #1190 was introduced where the config file used
was the sample that came from dotcloud/docker-registry instead of
using the confd templated config we use today. Updating the file
to upstream's changes with the common layer as well as changing
the reference to the file in the Dockerfile fixes this regression.