chore(*): update CoreOS to 402.2.0; Docker to 1.1.2
Note that this release is important because it has several big fixes:

* etcd 0.4.6 (with a fix for a major memory leak)
* fleet 0.6.2 (with major rewrites for stability)
* Docker 1.1.2 (with additional port allocation and container fixes)

Because with this Docker release the format of `docker inspect`
has changed, we must change the ANNOUNCE_TEMPLATE. We also have to
change `fleetctl list-units` in the CoreOS scheduler because we
are getting more columns of output now.

We also must change the device name on EC2 back to `/dev/sda`. See
https://github.com/coreos/scripts/pull/311