fix(docs): remove coveralls badge and mentions of coverage stats
The "95% passing" badge is stale since we don't use Travis-ci.org now,
and regardless was only summarizing coverage on the controller/
subproject. It would be great to roll up the coverage stats we do have
in the Go and Python subprojects and summarize that number in the
README, but for now we should hide this misleading number. Likewise,
the .travis.yml file was removed as it can only lead people astray
at this point.