feat(contrib/digitalocean): Use native CoreOS integration
Use the native CoreOS experience as recently added by DigitalOcean.

This rewrites the previous provisioning script and replaced the tugboat
dependency by the new docl tool dependency.
Docl was written specifically to support the operations we need.
Once a new DigitalOcean CLI comes out with support for the metadata
flag the use of Docl can be reexamined.