ref(builder): replace python with bash
Debugging the builder script when it fails is really difficult in
python. You have to import a bunch of libraries, pipe stderr on some
commands to stdout, inspect the return codes and the error returned...
In short, it's quite difficult to debug. Refactoring to bash allows us
to debug the script quite easily by setting the -x option for bash
scripts and we have instant debugging.

This refactor also removes the "remote:" prefix from most of the
output, which is another huge plus.