feat(start): wait on containers to start
Previously, containers would start and would be considered "running"
once Fleet launched the container and flagged the job as "launched".
The job was launched, but that doesn't mean that the container was
running. Dropping down to systemd's level (exposed through Fleet)
allows us to wait until the container is actually running. In the case
of data containers, it waits until they have exited.

This commit also removes the start logic from `deisctl install` so that
we can just load the unit files and configure the platform before
starting it.