fix(logspout): use custom log format
Go's syslog package hardcodes a lot of the log message format. This
is very difficult to customize for our own use. For example, we want
to follow a different timestamp format than RFC3339. We also want to
remove the hostname from the logs since we don't want users to "know"
where each container is being run from in the logs. In our case, though,
the hostname is the container ID for deis-logspout on that host which is
useless for the user, anyways. Connecting to the syslog server directly
allows us to format the message in a sensible manner.