fix(registry): make deis-cache mandatory
Sometimes the redis cache would take longer than the registry to come
up. In that case, confd would then set the host and port to be "",
in which case the registry would parse the YAML value as `None`. This
change makes the cache mandatory in order for the registry to come up.