fix(controller): properly serialize JSONField objects
JSONField is a finnicky thing. When serializing the object in the API,
it will just dump the object as a string. This ends in an invalid JSON
response. In the case of the Config object:

    {"owner": "bacongobbler", "app": "go", "values": "{\"HELLO\": \"world\"}", "memory": "{}", "cpu": "{}", "tags": "{}", ...}

This fix changes the JSON output to the following:

    {"owner": "bacongobbler", "app": "go", "values": {"HELLO": "world"}, "memory": {}, "cpu": {}, "tags": {}}

This is completely backwards compatible with previous requests. The
JSONSerializer will accept both strings and JSON objects in the request.