fix(builder): properly parse config vars
Previously, the entire export from `jq` was considered one string, so
when we printed the output it would be like "-e HELLO=foo GOODBYE=bar".
This fixes it to "-e HELLO=foo -e GOODBYE=bar".

I also removed the for loop in favour of a list as bash invokes a
subshell in for loops, losing any data you set (such as BUILD_OPTS).