fix(router): only add x-forwarded-proto on https
Prior to 1f74eeb, sending decrypted https traffic to port 80 worked fine
in the case that there was an https endpoint/load balancer in front of
the cluster. In this example, the addition of X-Forwarded-Proto changes
the forwarded protocol from https to http. Making the prototype only
available when an SSL certificate and key is installed fixes this.