fix(store): don't output secretKey and accessKey
In 33fb20e7f94afe164f2ceeb34fc5f2379934f3f7, the removal of the
output redirect for `radosgw-admin user create` means that the
accessKey and secretKey are output in the journal when the user
is created. As these are secrets, they shouldn't ever be printed.

This commit redirects the output of this command to /dev/null.