fix(deisctl): prevent build on windows
On windows the deisctl client will build, but it runs into a lot of
problems at runtime due to limitations in fleetctl. It depends on the
ssh-agent domain socket which cannot exist in windows as windows does
not support the idea of domain sockets.

This will prevent issues at runtime at the expense of being able to
build and run deisctl in windows.