fix(builder): revert to testing local package
`go test` doesn't like it if there's multiple main scripts in the same
directory. Since there are no other packages to test, `.` tests
everything.

From a platform-wide perspective this does seem odd, but it's only
because builder is not built with Go from the ground up, whereas the
logger, logspout and publisher are pure Go.