fix(store): list all monitors in [global] section of config
This is due to a Ceph bug - http://tracker.ceph.com/issues/10012.
Some configuration options aren't read properly if they're not in
the `[global]` section of the Ceph config. This means that the
store-gateway component was only connecting to the initial monitor,
instead of using all the enumerated monitors.

This notation is functionally equivalent, so it'll work just fine
for us.