feat(contrib/coreos): add etcd debug unit
This adds an optional service to aid in debugging etcd issues.
Note that it's only enabled in the cloud-config file, not started.
This enables users to `sudo systemctl start debug-etcd` on a machine
before performing tasks which may strain etcd (such as heavy IO),
and then watch the service's journal to see how things like heartbeat
intervals change.