fix(registry): use empty strings and not nil in config
In #1858, the LRU cache was activated for registry. This exposed
an issue with how config settings are handled by docker-registry -
specifically, it doesn't appear to handle the YAML nil (`~`) value
properly. In the config parser logic in docker-registry
(see: https://github.com/deis/docker-registry/blob/2d50c1c6f396ac68858135be05dbfe41fe1898e2/docker_registry/lib/config.py#L61),
values are expected to be strings, dicts, or non-existent. The default value
when an option exists but is unset is an empty string.
This commit changes the cache config settings in
registry to use an empty string instead of ~.

Additionally, this commit also ensures that deisctl starts cache
before registry to ensure the registry is using the LRU cache.

closes #2833