fix(contrib/ec2): prepare etcd before starting etcd
coreos-cloudinit processes units as they are listed in the cloud-config
file, so even though `prepare-etcd-data-directory.service` asks to be
started before etcd with `Before=etcd.service`, etcd has already
been started by the time this unit is encountered.

This commit prepends the Docker volume and etcd data volme preparation
to the main unit list, so they are started before etcd/fleet.

TESTING: Provision an EC2 cluster. You should notice that
`/media/ephemeral/etcd` has `log` and `conf` files, as well as a
`snapshots` directory. The lack of a snapshots directory is what
resulted in a panic when etcd tried to write a snapshot to disk.