ref(builder): Remove use of root in gitreceive
This change removes the use of root in gitreceive. Before we had to run
a script as root to get Docker access, but this has since been fixed in
Docker.

The motivation between the change was because of bug #1843, caused by
running "git gc" as root. This changed the owner of git files and
directories to root:root, which was inaccessable from git:git.

This also fixes /buildpacks check which checked if /buildpacks was a
file (which it is not, it's a directory).