fix(logger): use godep for dependencies
This issue was brought up when building a new release of images for
deployment. The logger relied on `go get` to fetch dependencies. Because
there was a breaking change in the etcd client[1], the logger failed to
connect to the cluster. We have moved over to a global Godeps for
managing dependencies, so this changes the compilation on the host to
follow the other components' build process.

[1]: https://github.com/coreos/go-etcd/commit/5348c80bde6223fa53c48ef546b4802db69b1d70#diff-33c1223aac4ef741731cc8c73b4d7777L295