fix(controller): destroy app containers in parallel
When we call app.destroy(), each container is destroyed in series. Since
we now rely on Fleet's state to respond, we are getting a more accurate
idea of what state the container is actually in. Because of this,
c.destroy() was taking a very long time waiting for a response from
fleet. iThis wasn't a problem before because we were waiting for
django-fsm, which basically said "yup, I told fleet to destroy the
container. It's dead now", which wasn't entirely true.

Switching to destroying containers in parallel makes this operation much
faster.