feat(client): add client extensions
If an unknown command is specified, the client will attempt to execute
the command as a dash-separated command. If the command cannot be found,
The original usage message is returned. This opens the opportunity to
add and install extensions to the deis client by simply installing a
plugin onto the operating system.