feat(contrib/ec2/gen-json.py): split private and public subnets
When installing a Deis cluster in an existing VPC, it is currently only
possible to specify one set of (public) subnets in which both the
Deis EC2 instances and the ELB are placed. It is not possible to
enhance security by placing the Deis hosts in a private section of the
VPC, into subnets that have no direct connection to the Internet.

This change adds the environment variable `VPC_PRIVATE_SUBNETS` to help
set up a Deis cluster in a 'split' VPC environment.