fix(router): Unify timeout values
The router was having different timeout values for keepalive_timeout
and for proxy timeouts.

This can cause issues with front facing load balancers in front of
deis-router (like Amazon ELB), and could be the cause of issue #3448

Timeout values should be greater than the timeout value of the front facing
load balancer. Ref:

> The keepalive timeout value on your backend server must be higher than that
> of your ELB connection timeout. If it is lower, the ELB will re-use the idle
> connection when your server has already dropped the connection, resulting in
> the client being served up a blank response.
http://engineering.chartbeat.com/2014/02/12/part-2-lessons-learned-tuning-tcp-and-nginx-in-ec2/

> The ELB timeout should be set to less then the backend server idle timeout.
https://forums.aws.amazon.com/thread.jspa?threadID=104688