chore(builder): migrate to heroku's cedar stack
Heroku now officially maintains a Docker image for their cedar stack.
The maintainer of cedarish is deprecating it in favour of this image,
so we should migrate over such that we emulate their environment
perfectly.

Since we no longer build a root filesystem and package it into a tarball,
this speeds up build time of the builder dramatically with the tradeoff
of having to fetch heroku/cedar:14 on boot, which is slightly slower
than a `docker import`. However, the difference between relying on your
local connection to build and the cloud provider using precious clock
cycles to build the image vs. relying on a cloud provider's bandwidth to
pull the image on boot is quite significant, so really it's a question
of tomato vs. tomatoe.