fix(store): lower number of placement groups
We were incorrectly configuring a large number of placement groups
for the Ceph storage pools, resulting in each store-daemon doing
more work than necessary.

This commit removes the logic to reconfigure the number of placement
groups for the default pools (since they are already larger than 64),
and lowers the default number of placement groups from 128 to 64,
resulting in 768 placement groups per daemon. We also reconfigure
the warning for number of placement groups per OSD to 800 so we
no longer warn in the typical case.