fix(store): fix shared etcd key defaults
In store-monitor's bin/boot script, we set default etcd keys that
are used to template ceph.conf for all store components. Previously,
this logic was inside the monitor setup lock logic, meaning it's only
executed once for the lifetime of a cluster. This breaks when adding new
keys to ceph.conf and this logic, since upgraded clusters will not
execute this logic a second time, thus skipping the creation of new
(and necessary) keys. The cluster will not come up on an upgrade
because confd cannot template ceph.conf in the store components.

Because *all* store monitors now set these defaults before checking
if they're the master, we have to ensure that all monitors are setting
the same default values and not writing over a different default set
by a competing monitor. We do this by removing the ability to set
environment variable overrides for these values. It never really made
sense anyway, since a user can always override these by manually setting
the etcd key before cluster start.