fix(*): docker.service depends on its mount
Just specifying that the mount needs to be started before Docker is
insufficient - since Docker is socket activated, it has a chance
to start before the mount. This commit adds a new drop-in unit
which configures Docker to only start after its mount on providers
where we use a separate mount for Docker (AWS, Azure, and GCE).