fix(builder): generate key on boot
This change generates keys on boot, as well as storing the keys in etcd
such that the key will persist between builder reboots.

BREAKING CHANGE: since the key is no longer generated on `docker build`,
the first time the user runs this they may receive a key conflict
warning with the old builder keys. Since this change now saves it to
etcd, the user should remove the old key from their known_hosts file
and accept the new key. On reboots, the newly generated key should stay
the same.